# web/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LogPage": {
        "description": "一页日志（按 id 升序）。\n\n:param next_since_id: 下一页的游标（本页为空时回传请求里的 `since_id`，\n    于是\"没有新日志\"不会把前端的游标清成 `None`）。",
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "logs": {
            "items": {
              "$ref": "#/components/schemas/LogRow"
            },
            "title": "Logs",
            "type": "array"
          },
          "next_since_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Since Id"
          }
        },
        "required": [
          "logs",
          "next_since_id",
          "limit"
        ],
        "title": "LogPage",
        "type": "object"
      },
      "LogRow": {
        "description": "`system_logs` 一行的 API 视图（WS 快照与 REST 分页共用同一形状）。",
        "properties": {
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "seq_in_task": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seq In Task"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "ts": {
            "title": "Ts",
            "type": "string"
          },
          "unit_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Ref"
          },
          "worker_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker Id"
          }
        },
        "required": [
          "id",
          "ts",
          "level",
          "source",
          "message"
        ],
        "title": "LogRow",
        "type": "object"
      },
...
      "PolicyRequest": {
        "additionalProperties": false,
        "description": "切自动放行策略（`grade_ab` = 一键全自动；`grade_a` = 一键回退）。",
        "properties": {
          "policy": {
            "enum": [
              "off",
              "grade_a",
              "grade_ab"
            ],
            "title": "Policy",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "policy"
        ],
        "title": "PolicyRequest",
        "type": "object"
      },
      "PolicyResponse": {
        "description": "策略切换的结论（`config_path` 让\"到底写到哪了\"可查）。",
        "properties": {
          "changed": {
            "title": "Changed",
            "type": "boolean"
          },
          "config_path": {
            "title": "Config Path",
            "type": "string"
          },
          "policy": {
            "title": "Policy",
            "type": "string"
          },
          "previous": {
            "title": "Previous",
            "type": "string"
          }
        },
        "required": [
          "policy",
          "previous",
          "changed",
          "config_path"
        ],
        "title": "PolicyResponse",
        "type": "object"
      },
...
      "PoolPauseRequest": {
        "additionalProperties": false,
        "description": "暂停 / 恢复一个池（`paused` 是**目标状态**，不是\"切换\"）。",
        "properties": {
          "paused": {
            "title": "Paused",
            "type": "boolean"
          },
          "pool": {
            "enum": [
              "draft",
              "voice",
              "render",
              "publish"
            ],
            "title": "Pool",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "pool",
          "paused"
        ],
        "title": "PoolPauseRequest",
        "type": "object"
      },
      "PoolPauseResponse": {
        "description": "暂停 / 恢复的结论（`note` 直接给人看：\"在途 N 个跑完\"）。",
        "properties": {
          "changed": {
            "title": "Changed",
            "type": "boolean"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "paused": {
            "title": "Paused",
            "type": "boolean"
          },
          "paused_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paused At"
          },
          "paused_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paused By"
          },
          "pending": {
            "title": "Pending",
            "type": "integer"
          },
          "pool": {
            "title": "Pool",
            "type": "string"
          },
          "running": {
            "title": "Running",
            "type": "integer"
          }
        },
        "required": [
          "pool",
          "paused",
          "changed",
          "running",
          "pending",
          "note"
        ],
        "title": "PoolPauseResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AI 全自动营销号制片台",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/logs": {
      "get": {
        "description": "`since_id` 给定 ⇒ 增量；不给 ⇒ 最近 `limit` 条（都按 id 升序返回）。",
        "operationId": "list_logs_api_v1_logs_get",
        "parameters": [
          {
            "description": "只取 id 大于它的日志（增量）",
            "in": "query",
            "name": "since_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "只取 id 大于它的日志（增量）",
              "title": "Since Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "级别下限：debug|info|warn|error|fatal",
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "级别下限：debug|info|warn|error|fatal",
              "title": "Level"
            }
          },
          {
            "description": "只看某个任务",
            "in": "query",
            "name": "task_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "只看某个任务",
              "title": "Task Id"
            }
          },
          {
            "description": "只看某个来源，如 pool.voice",
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "只看某个来源，如 pool.voice",
              "title": "Source"
            }
          },
          {
            "description": "只取 id 小于等于它的日志（往前翻历史）",
            "in": "query",
            "name": "until_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "只取 id 小于等于它的日志（往前翻历史）",
              "title": "Until Id"
            }
          },
          {
            "description": "message / source 的字面子串",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "message / source 的字面子串",
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Logs",
        "tags": [
          "logs"
        ]
      }
    },
    "/api/v1/logs/export": {
      "get": {
        "description": "导出 NDJSON（一行一条日志，字段与 `/api/v1/logs` 完全一致）。\n\n行数**恰等于 `limit`** ⇒ 大概率被截断（前端据此提示\"收紧过滤条件再导\"）；\n服务端不额外回一行 meta —— 那会让这个文件不再是干净的 NDJSON。",
        "operationId": "export_logs_api_v1_logs_export_get",
        "parameters": [
          {
            "description": "从这条之后开始导出（增量）",
            "in": "query",
            "name": "since_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "从这条之后开始导出（增量）",
              "title": "Since Id"
            }
          },
          {
            "description": "导出到这条为止（含）",
            "in": "query",
            "name": "until_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "导出到这条为止（含）",
              "title": "Until Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20000,
              "maximum": 200000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "级别下限",
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "级别下限",
              "title": "Level"
            }
          },
          {
            "description": "只看某个任务",
            "in": "query",
            "name": "task_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "只看某个任务",
              "title": "Task Id"
            }
          },
          {
            "description": "只看某个来源",
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "只看某个来源",
              "title": "Source"
            }
          },
          {
            "description": "message / source 的字面子串",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "message / source 的字面子串",
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Logs",
        "tags": [
          "logs"
        ]
      }
    },
...
    "/api/v1/overview/auto": {
      "post": {
        "description": "切自动放行策略（写 `config/app.yaml` + 同步内存 + `audit_ops` · 裁定 139）。",
        "operationId": "set_auto_approve_policy_api_v1_overview_auto_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Auto Approve Policy",
        "tags": [
          "overview"
        ]
      }
    },
    "/api/v1/overview/pools": {
      "post": {
        "description": "暂停 / 恢复一个池（**在途跑完**，不重启 worker · 裁定 138）。",
        "operationId": "set_pool_paused_api_v1_overview_pools_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PoolPauseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoolPauseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Pool Paused",
        "tags": [
          "overview"
        ]
      }
    },
...
  }
}